[Improve][Connector-V2] Refactor the package of hdfs file connector (#2402)
