+++ seatunnel-connectors-v2/connector-file/connector-file-hadoop/src/main/java/org/apache/seatunnel/connectors/seatunnel/file/hdfs/sink/filesystem/HdfsFileSystem.java
@@ -15,18 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.connectors.seatunnel.file.sink.hdfs;
+package org.apache.seatunnel.connectors.seatunnel.file.hdfs.sink.filesystem;
 
-import org.apache.seatunnel.api.sink.SeaTunnelSink;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.AbstractFileSink;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.spi.SinkFileSystemPlugin;
+import org.apache.seatunnel.connectors.seatunnel.file.hdfs.sink.util.HdfsUtils;
+import org.apache.seatunnel.connectors.seatunnel.file.sink.spi.FileSystem;
 
-import com.google.auto.service.AutoService;
+import org.apache.hadoop.fs.Path;
 
-@AutoService(SeaTunnelSink.class)
-public class HdfsFileSink extends AbstractFileSink {
+import java.io.IOException;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class HdfsFileSystem implements FileSystem {
     @Override
-    public SinkFileSystemPlugin getSinkFileSystemPlugin() {
-        return new HdfsFileSinkPlugin();
+    public void deleteFile(String path) throws IOException {
+        HdfsUtils.deleteFile(path);
+    }
+
+    @Override
+    public List<String> dirList(String dirPath) throws IOException {
+        List<Path> paths = HdfsUtils.dirList(dirPath);
+        return paths.stream().map(dir -> dir.getName()).collect(Collectors.toList());
     }
 }
